pyproject.toml
[project]
name = "agent-projects"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=5.46.1",
    "openai>=1.108.1",
    "pypdf>=6.1.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
